+++ lib/vonage/client.rb
@@ -103,6 +103,13 @@ module Vonage
       @pricing ||= T.let(PricingTypes.new(config), T.nilable(Vonage::PricingTypes))
     end
 
+    # @return [ProactiveConnect]
+    #
+    sig { returns(T.nilable(Vonage::ProactiveConnect)) }
+    def proactive_connect
+      @proactive_connect ||= T.let(ProactiveConnect.new(config), T.nilable(Vonage::ProactiveConnect))
+    end
+
     # @return [Redact]
     #
     sig { returns(T.nilable(Vonage::Redact)) }
@@ -124,6 +131,13 @@ module Vonage
       @sms ||= T.let(SMS.new(config), T.nilable(Vonage::SMS))
     end
 
+    # @return [Subaccounts]
+    #
+    sig { returns(T.nilable(Vonage::Subaccounts)) }
+    def subaccounts
+      @subaccounts ||= T.let(Subaccounts.new(config), T.nilable(Vonage::Subaccounts))
+    end
+
     # @return [TFA]
     #
     sig { returns(T.nilable(Vonage::TFA)) }
@@ -136,6 +150,13 @@ module Vonage
     sig { returns(T.nilable(Vonage::Verify)) }
     def verify
       @verify ||= T.let(Verify.new(config), T.nilable(Vonage::Verify))
+    end
+
+    # @return [Verify2]
+    #
+    sig { returns(T.nilable(Vonage::Verify2)) }
+    def verify2
+      @verify2 ||= T.let(Verify2.new(config), T.nilable(Vonage::Verify2))
     end
 
     # @return [Voice]