+++ lib/vonage/proactive_connect/lists/list_response.rb
@@ -1,0 +1,11 @@
+# typed: true
+
+class Vonage::ProactiveConnect::Lists::ListResponse < Vonage::Response
+  include Enumerable
+
+  def each
+    return enum_for(:each) unless block_given?
+
+    @entity._embedded.lists.each { |item| yield item }
+  end
+end